frontend/requirements.txt
streamlit
pyrebase4
requests
pytest
dotenv